Move params unit test to params_test.go
